change behavior to detect config files.

+++ src/main/kotlin/co/anbora/labs/sqlfluff/ide/startup/InitConfigFiles.kt
@@ -42,10 +42,10 @@ class InitConfigFiles: ProjectActivity {
 
     private fun checkDefaultConfigFile() {
         val pathConfig = DEFAULT_CONFIG_PATH
-        //if (!pathConfig.exists()) {
+        if (!pathConfig.exists()) {
             val configFile = pathConfig.toFile().createNewFile()
             Files.write(pathConfig, LANGUAGE_DEMO_TEXT.toByteArray())
-        //}
+        }
     }
 
     companion object {
